# openapi.json
{"openapi": "3.1.0", "info": {"title": "Little Research Lab API", "version": "0.1.0"}, "paths": {"/api/auth/login": {"post": {"tags": ["Auth"], "summary": "Login For Access Token", "operationId": "login_for_access_token_api_auth_login_post", "requestBody": {"content": {"application/x-www-form-urlencoded": {"schema": {"$ref": "#/components/schemas/Body_login_for_access_token_api_auth_login_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Token"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/auth/logout": {"post": {"tags": ["Auth"], "summary": "Logout", "operationId": "logout_api_auth_logout_post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/auth/me": {"get": {"tags": ["Auth"], "summary": "Read Users Me", "operationId": "read_users_me_api_auth_me_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/content": {"get": {"tags": ["Content"], "summary": "List Content", "operationId": "list_content_api_content_get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "status", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Status"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/ContentItemResponse"}, "title": "Response List Content Api Content Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["Content"], "summary": "Create Content", "operationId": "create_content_api_content_post", "security": [{"OAuth2PasswordBearer": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ContentCreateRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ContentItemResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/content/{item_id}": {"get": {"tags": ["Content"], "summary": "Get Content", "operationId": "get_content_api_content__item_id__get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "item_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Item Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ContentItemResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["Content"], "summary": "Update Content", "operationId": "update_content_api_content__item_id__put", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "item_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Item Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ContentUpdateRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ContentItemResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["Content"], "summary": "Delete Content", "operationId": "delete_content_api_content__item_id__delete", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "item_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Item Id"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/assets": {"get": {"tags": ["Assets"], "summary": "List Assets", "operationId": "list_assets_api_assets_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/AssetResponse"}, "type": "array", "title": "Response List Assets Api Assets Get"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}, "post": {"tags": ["Assets"], "summary": "Upload Asset", "operationId": "upload_asset_api_assets_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_upload_asset_api_assets_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AssetResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/assets/{asset_id}/content": {"get": {"tags": ["Assets"], "summary": "Get Asset Content", "operationId": "get_asset_content_api_assets__asset_id__content_get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "asset_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Asset Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/users": {"get": {"tags": ["Users"], "summary": "List Users", "operationId": "list_users_api_users_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/UserResponse"}, "type": "array", "title": "Response List Users Api Users Get"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}, "post": {"tags": ["Users"], "summary": "Create User", "operationId": "create_user_api_users_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserCreateRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/users/{user_id}": {"put": {"tags": ["Users"], "summary": "Update User", "operationId": "update_user_api_users__user_id__put", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "title": "User Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserUpdateRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/public/home": {"get": {"tags": ["Public"], "summary": "Get Public Home", "operationId": "get_public_home_api_public_home_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/public/content/{slug}": {"get": {"tags": ["Public"], "summary": "Get Public Content", "operationId": "get_public_content_api_public_content__slug__get", "parameters": [{"name": "slug", "in": "path", "required": true, "schema": {"type": "string", "title": "Slug"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ContentItemResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/health": {"get": {"summary": "Health Check", "operationId": "health_check_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"AssetResponse": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "filename_original": {"type": "string", "title": "Filename Original"}, "mime_type": {"type": "string", "title": "Mime Type"}, "size_bytes": {"type": "integer", "title": "Size Bytes"}, "visibility": {"type": "string", "enum": ["public", "unlisted", "private"], "title": "Visibility"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["id", "filename_original", "mime_type", "size_bytes", "visibility", "created_at"], "title": "AssetResponse"}, "Body_login_for_access_token_api_auth_login_post": {"properties": {"json_data": {"anyOf": [{"$ref": "#/components/schemas/LoginRequest"}, {"type": "null"}]}, "grant_type": {"anyOf": [{"type": "string", "pattern": "^password$"}, {"type": "null"}], "title": "Grant Type"}, "username": {"type": "string", "title": "Username"}, "password": {"type": "string", "format": "password", "title": "Password"}, "scope": {"type": "string", "title": "Scope", "default": ""}, "client_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Client Id"}, "client_secret": {"anyOf": [{"type": "string"}, {"type": "null"}], "format": "password", "title": "Client Secret"}}, "type": "object", "required": ["username", "password"], "title": "Body_login_for_access_token_api_auth_login_post"}, "Body_upload_asset_api_assets_post": {"properties": {"file": {"type": "string", "format": "binary", "title": "File"}}, "type": "object", "required": ["file"], "title": "Body_upload_asset_api_assets_post"}, "ContentBlockModel": {"properties": {"id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Id"}, "block_type": {"type": "string", "enum": ["markdown", "image", "chart", "embed", "divider"], "title": "Block Type"}, "data_json": {"additionalProperties": true, "type": "object", "title": "Data Json"}, "position": {"type": "integer", "title": "Position"}}, "type": "object", "required": ["block_type", "data_json", "position"], "title": "ContentBlockModel"}, "ContentCreateRequest": {"properties": {"title": {"type": "string", "title": "Title"}, "slug": {"type": "string", "title": "Slug"}, "summary": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Summary"}, "status": {"type": "string", "enum": ["draft", "scheduled", "published", "archived"], "title": "Status", "default": "draft"}, "visibility": {"type": "string", "enum": ["public", "unlisted", "private"], "title": "Visibility", "default": "public"}, "publish_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Publish At"}, "type": {"type": "string", "enum": ["post", "page"], "title": "Type"}, "blocks": {"items": {"$ref": "#/components/schemas/ContentBlockModel"}, "type": "array", "title": "Blocks", "default": []}}, "type": "object", "required": ["title", "slug", "type"], "title": "ContentCreateRequest"}, "ContentItemResponse": {"properties": {"title": {"type": "string", "title": "Title"}, "slug": {"type": "string", "title": "Slug"}, "summary": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Summary"}, "status": {"type": "string", "enum": ["draft", "scheduled", "published", "archived"], "title": "Status", "default": "draft"}, "visibility": {"type": "string", "enum": ["public", "unlisted", "private"], "title": "Visibility", "default": "public"}, "publish_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Publish At"}, "id": {"type": "string", "format": "uuid", "title": "Id"}, "type": {"type": "string", "enum": ["post", "page"], "title": "Type"}, "published_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Published At"}, "owner_user_id": {"type": "string", "format": "uuid", "title": "Owner User Id"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}, "blocks": {"items": {"$ref": "#/components/schemas/ContentBlockModel"}, "type": "array", "title": "Blocks", "default": []}}, "type": "object", "required": ["title", "slug", "id", "type", "owner_user_id", "created_at", "updated_at"], "title": "ContentItemResponse"}, "ContentUpdateRequest": {"properties": {"title": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Title"}, "slug": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Slug"}, "summary": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Summary"}, "status": {"anyOf": [{"type": "string", "enum": ["draft", "scheduled", "published", "archived"]}, {"type": "null"}], "title": "Status"}, "visibility": {"anyOf": [{"type": "string", "enum": ["public", "unlisted", "private"]}, {"type": "null"}], "title": "Visibility"}, "publish_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Publish At"}, "blocks": {"anyOf": [{"items": {"$ref": "#/components/schemas/ContentBlockModel"}, "type": "array"}, {"type": "null"}], "title": "Blocks"}}, "type": "object", "title": "ContentUpdateRequest"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "LoginRequest": {"properties": {"username": {"type": "string", "title": "Username"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["username", "password"], "title": "LoginRequest"}, "Token": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "token_type": {"type": "string", "title": "Token Type"}}, "type": "object", "required": ["access_token", "token_type"], "title": "Token"}, "UserCreateRequest": {"properties": {"email": {"type": "string", "title": "Email"}, "display_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Display Name"}, "password": {"type": "string", "title": "Password"}, "roles": {"items": {"type": "string"}, "type": "array", "title": "Roles", "default": ["editor"]}}, "type": "object", "required": ["email", "password"], "title": "UserCreateRequest"}, "UserResponse": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "email": {"type": "string", "title": "Email"}, "display_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Display Name"}, "roles": {"items": {"type": "string"}, "type": "array", "title": "Roles", "default": []}, "status": {"type": "string", "title": "Status"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["id", "email", "status", "created_at"], "title": "UserResponse"}, "UserUpdateRequest": {"properties": {"roles": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Roles"}, "status": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Status"}}, "type": "object", "title": "UserUpdateRequest"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"OAuth2PasswordBearer": {"type": "oauth2", "flows": {"password": {"scopes": {}, "tokenUrl": "/api/auth/login"}}}}}}